+++ backend/common/sql_models/group_vdbs.py
@@ -23,6 +23,7 @@ class GroupVDBs(Base):
     gvdbs_url: Mapped[str] = mapped_column(VARCHAR, nullable=False)
     gvdbs_collection: Mapped[str] = mapped_column(VARCHAR, nullable=False)
     gvdbs_retr_params: Mapped[str] = mapped_column(VARCHAR, nullable=False)
+    gvdbs_retr_filters: Mapped[str] = mapped_column(VARCHAR, nullable=False, server_default=text("'{}'"))
     gvdbs_emb_model: Mapped[str] = mapped_column(VARCHAR, nullable=False)
     gvdbs_created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     # gvdbs_status = success / warning / danger